PServer/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
httpx
grpcio
grpcio-tools